Split HandshakeProtocol into another class
